Added a table component with a corresponding css file.  Added a game.css file to go with the Game component.  Floated the Table to the right inside the game component.
Refactored PlayerCards out of Game component.  Created a Seats component to represent the players seated around the table.  Added in the Player Components into the Seats components.

Table is floating right, filling 75% of the parent div.

Added die component and added six of them to the Table Component.

Created HumanPlayer and ComputerPlayer components and refactored PlayerCard to render those components dependent on whether or not the isHuman prop is true or false.

Refactored roll method to preRoll so we can use a roll command to have a regular dice roll for the game.

Updated preRoll so that once a player rolls, state for that player is set so rollAgain = false.

Enabled Human player to make their first roll once the game has started.

roll action in redux gameActions rolls 6 initial dice and calls the reducer which updates the state of the dice.

Initial roll now changes the dice to their proper ui picture.

Updated gameReducer initial state test to passing.  Updated PlayerCard test to test that the right component is showing, ComputerPlayer or HumanPlayer based on the isHuman prop value.

Removed extra line space in PlayerCard test.