Removed console logs.  Added new function getDiceNotHeld, that takes in all six dice and returns only the dice that aren't being held.
Added Redux action to determine whether or not a player is able to end their turn or not, and fires a reducer, not yet written, to update the state.

Updated CHANGE_TURN reducer so it also enables the roll of the players turn it is changing to.

Wrote two tests for canEndTurn redux Action.