Created PlayByPlay Component and put component imside Table component.  Removed two console logs from gameActions redux action.
Updated PlayByPlay comnponent so it takes in turn, dice, and currentRollScore from state and maps it to props.  The component displays the players turn on screen.

Added output for dice roll.

currentRollScore state was not mapped to props in Game component causing an Nan value down the line.  I added mapped currentRollScore to props in that component.

Added a roll cannot be enabled conditional to the componentDidUpdate lifecycle method in the ComputerPlayer component in order for the computer to determine there next move.

Game flow improved.  Added new state item, checkedForPeezda, added a redux action to mark it as true, and added to the ROLL DICE reducers to flag the item as false, when new roll commences.

updated game reducer test so that it now tests the initial state with the new checkedForPeezda item.