newGameForm has been updated to allow for the user to choose a player type for each player they choose to have in the game.
For each player there is now an option to choose between human and computer in the new game form.

newGameForm is now setup to give an option for each player as to whether or not the player is human or computer.

newGameForm now correctly updates the component State when the number of players are change, and the player Type is changed.

Reran npm install to update package-lock.json.